Reorganize the utility functions
Separate the functions for finding libgmt.
